test: split file mention resource regressions
